Rewrite fcnt14 test
This test has been changed in order to generate testcases at runtime.

The new algorithm sets a random starting position on file using
lseek(), it randomly generates fcntl() parameters for parent and
child and it verifies that fcntl() will raise a blocking error on
child when it's supposed to.

The runtest file has been modified so we can run test with or without
mandatory locking using the new -l flag.