Add sphinx configuration
The sphinx configuration is the first step to deploy a new LTP website.
The way we do it, is to provide a configuration that contains a basic
index page plus reference to README and markdown documentation.

The myst_parser will take care to convert markdown documentation into a
proper website and renku theme will provide a pretty good looking
webpage theme.

To build the website, please use the following command:

	cd site/ && sphinx -b html . build

We will find our LTP website inside the site/build/ folder, under the
index.html file.