Rewrite openposix clock_settime bugged tests inside LTP
Both clock_settime_7-1 and clock_settime_8-1 are affected by the
same bug which I think it's related to the way tests are written.

sleep() is used by the parent and is not reliable. It can oversleep
or undersleep according to system overload or signals which are
received. And we never check for its return value. Using
clock_nanosleep would make parent more reliable in this case.

At the same time, the test is taking for granted a certain
synchronization between child and parent, which is not always true
in case of system overload.

My suggestion is to rewrite the test using LTP, which has better
timing handling (see tst_timer.h).

To: ltp@lists.linux.it
Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250620-clock_nanosleep05-60b3cfba52fe",
    "prefixes": []
  }
}